pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "instruction-video-generator"
version = "0.1.0"
description = "AI-powered instructional video generator using browser-use + TTS + ffmpeg"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "browser-use>=0.12.1",
  "browser-use-sdk>=0.3.7",
  "fastapi>=0.116.1",
  "imageio[ffmpeg]>=2.37.2",
  "numpy>=2.4.1",
  "openai>=2.16.0",
  "python-dotenv>=1.0.0",
  "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0.0",
]

[project.scripts]
instruction-video-generator = "instruction_video_generator.cli:main"
instruction-video-web = "instruction_video_generator.web:main"

[tool.setuptools.packages.find]
include = ["instruction_video_generator*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
